backend/requirements.txt
django
daphne
psycopg2-binary
djangorestframework
djangorestframework-simplejwt
channels
Pillow
django-cors-headers